refactor(storage): extract SQLiteDeletionMixin from sqlite _base.py
Peel the coupled retention-hooks + chunked bulk-delete cluster out of
SQLiteStorageBase into a new SQLiteDeletionMixin (base/_deletion.py),
composed into SQLiteStorage ahead of SQLiteStorageBase so the atomic
_retention_perform_delete override wins over RetentionMixin's default.

Method bodies are byte-identical (AST-verified); only the @handle_exceptions
decorator line is rewritten to @SQLiteStorageBase.handle_exceptions and the
travelling imports (RetentionTarget, chunked, RETENTION_DELETE_CHUNK) moved
with the methods. BootSchema, the module-level utils, connection primitives,
and clear_user_data stay residual (clear_user_data reaches the moved helpers
via MRO). No public-surface or behavior change.
