+++ reflexio/server/services/storage/sqlite_storage/base/__init__.py
@@ -1,0 +1,5 @@
+"""SQLite storage base sub-mixins (peeled from _base.py)."""
+
+from ._deletion import SQLiteDeletionMixin
+
+__all__ = ["SQLiteDeletionMixin"]
